Block Breakers: Don't try to break liquids. Closes #14.

+++ src/main/java/fi/dy/masa/autoverse/tileentity/TileEntityBlockBreaker.java
@@ -223,7 +223,9 @@ public class TileEntityBlockBreaker extends TileEntityAutoverseInventory
         {
             IBlockState state = world.getBlockState(pos);
 
-            if (state.getBlock().isAir(state, world, pos) == false && state.getBlockHardness(world, pos) >= 0f)
+            if (state.getBlock().isAir(state, world, pos) == false &&
+                state.getMaterial().isLiquid() == false &&
+                state.getBlockHardness(world, pos) >= 0f)
             {
                 world.destroyBlock(pos, true);
             }
